mi; fix output for breakpoints with multiple locations
Currently, the MI output for breakpoints with multiple locations is
broken.  It doesn't respect the MI output syntax, and is not really
sensible.  Here is an example with the output of -break-info, but it
also affects the output of -break-insert.

^done,
  BreakpointTable={
    nr_rows="2",
    nr_cols="6",
    hdr=[
      {width="7",alignment="-1",col_name="number",colhdr="Num"},
      ...
    ],
    body=[
      bkpt={
	number="1",
	...,
	addr="<MULTIPLE>",
	...,
      },
      {
	number="1.1",
	...,
      },
      {
	number="1.2",
	...,
      },
      bkpt={
	number="2",
	...,
	addr="<MULTIPLE>",
	...,
      },
      {
	number="2.1",
	...,
      },
      {
	number="2.2",
	...,
      }
    ]
  }

As you can see, the locations just follow the main breakpoint in the
list, in a way that does not respect MI output syntax and is not very
practical.  I think the natural way would be to have a list of locations
inside each breakpoint, such as:

^done,
  BreakpointTable={
    nr_rows="2",
    nr_cols="6",
    hdr=[
      {width="7",alignment="-1",col_name="number",colhdr="Num"},
      ...,
    ],
    body=[
      bkpt={
	number="1",
	...,
	addr="<MULTIPLE>",
	...,
	locations=[
	  {
	    number="1.1",
	    ...,
	  },
	  {
	    number="1.2",
	    ...,
	  }
	]
      },
      bkpt={
	number="2",
	...,
	addr="<MULTIPLE>",
	...,
	locations=[
	  {
	    number="2.1",
	    ...,
	  },
	  {
	    number="2.2",
	    ...,
	  }
	]
      }
    ]
  }

Now, I understand this represents a change in a published interface.
However, the output for multi-locations has never been valid so far.  I
don't know about all the frontends, but we realized that Eclipse
recognized the output was garbage and was ignoring most of the data.  So
in that case, it would be benificial to fix the bug.  I don't know if
there are other frontends out there that do expect the invalid output.

gdb/ChangeLog:

	* breakpoint.c (print_one_breakpoint): Print locations inside
	breakpoint for MI-like uiouts.