Fix threads left stopped after fork+thread spawn (PR threads/18600)
I am posting the fix suggested by Pedro for bug 18600 [1, 2].  I have to
admit that I don't fully understand the fix, but it fixes the bug and
causes no regression according in the testsuite.

The first issue is that threads were left stopped while gdb was thinking
they were running.  The second issue, uncovered after fixing the first
one, was that exited inferiors were left when they should have been
removed.

I tried to make a test for this situation, but it's been a bit more
difficult than I expected.  The idea is that the inferior forks a
certain number of times and waits for all children to exit.  Each fork
child spawns a number of threads that do nothing and joins them
immediately.  Normally, the program should run unimpeded (from the point
of view of the user) and exit very quickly.  Without this fix, it
doesn't because of some threads left stopped by gdb.  My only problem is
that the prompt comes back as soon as any inferior (even a child) exits,
and that moment is too early to say if the test passed or not.  I had to
resort to make it sleep for a second and then check that no thread is
left.  If you have suggestions on how to make the test more robust, they
are very welcome.

If this is acceptable, I would suggest backporting the fix to the 7.10
branch (it's currently in the TODO on the 7.10 wiki page).

gdb/ChangeLog:

	* linux-nat.c (linux_handle_extended_wait): On CLONE event,
	always mark the new thread as resumed.  Remove STOPPING
	parameter.
	(wait_lwp): Report to the core when thread group leader exits.
	Adjust call to linux_handle_extended_wait.
	(linux_nat_filter_event): Adjust call to
	linux_handle_extended_wait.
	(resume_stopped_resumed_lwps): Add debug output.

gdb/testsuite/ChangeLog:

	* gdb.threads/fork-plus-threads.c: New file.
	* gdb.threads/fork-plus-threads.exp: New file.

[1] https://sourceware.org/ml/gdb-patches/2015-07/msg00186.html
[2] https://sourceware.org/ml/gdb-patches/2015-07/msg00190.html