Introduce dprintf-flush-function for call-style dprintfs
A little bit of context: a colleague of mine likes debugging the JVM and
wants to use dynamic printfs to do so.  Unfortunately, openjdk sets the
standard output as fully buffered, as opposed to line buffered by
default.  This means that when using simple dprintfs such as:

  (gdb) dprintf file.c:1234, "I'm here\n"

the output won't show up on the terminal.  He can always stop the
program and manually call fflush from gdb.  Not very practical for
debugging.

He could also call setlinebuf(stdout) manually.  However, this can't be
done before the inferior is started, and it's possible for the inferior
to change it back.  He could also write a wrapper to printf that calls
fflush after printing.  In both cases you need some quite good insight
to guess why your dprintfs are not showing up.  I'd rather have something
that "just works".

The parameter dprintf-flush-function is meant to refer to a function
called after dprintf-function was called, in order to do the flushing.
By default, it is set to fflush stdout, which makes sense since the
default value of dprintf-function (printf) prints on stdout.  If
dprintf-channel is set, it passes that value instead to the function.

A test is included.  It doesn't do much, it just verifies that the
specified flush function indeed gets called.  I also updated dprintf.exp
to test the error message when dprintf-function is set to an empty
value.

gdb/ChangeLog:

	* breakpoint.c (update_dprintf_command_list): Add call to the
	flush function to the command list, factor out code.
	(_initialize_breakpoint): Register set/show
	dprintf-flush-function.
	(dprintf_flush_function): New global.
	(get_dprintf_call_print_line): New function.
	(get_dprintf_call_flush_line): New function.
	* NEWS: Mention set/show dprintf-flush-function.

gdb/testsuite/ChangeLog:

	* gdb.base/dprintf-flush.exp: New file.
	* gdb.base/dprintf-flush.c: New file.
	* gdb.base/dprintf.exp: Test error message when dprintf-function
	is empty.

gdb/doc/ChangeLog:

	* gdb.texinfo (Dynamic Printf): Document set
	dprintf-flush-function.