trace-support.exp: Add gdb_arch_supports_fast_trace and use it
Factor out this code, so we can use it at other places.  It would be
better to find a way to query gdb to know if the current architecture
supports fast tracepoints, instead of hard-coding a list of architectures
here.  But at least, we'll only have to change it at a single place if we do
that in the future.

Also, move the check to where test_fast_tracepoints is called, which
removes an indentation level for a lot of code.

gdb/testsuite/ChangeLog:

	* gdb.trace/ftrace.exp (test_fast_tracepoints): Move target
	support check to call site.  Factor out code to...
	* lib/trace-support.exp (gdb_target_supports_fast_trace): ...this.