gdbserver: pass pointer to struct tracepoint to install_fast_tracepoint_jump_pad
We currently pass many values to install_fast_tracepoint_jump_pad that
come from fields of struct tracepoint, and we plan to add another (the
kind field).  It would be much simpler to just pass a pointer to the
tracepoint and have install_fast_tracepoint_jump_pad fetch the fields.

That requires moving the struct tracepoint definition to tracepoint.h.