Support tracepoints ARM linux in GDBServer.
This patch adds support for tracepoints for ARM linux in GDBServer.

To enable this, this patch also introduces a new :K (kind) field in the QTDP
packet to encode the breakpoint kind, this is the same kind as a z0 packet.

This is the new qSupported feature : TracepointKinds

This field is decoded by breakpoint_from_length target ops in linux-low.

A note about tests :

Some tests expect to use a frame pointer register, however since there no such
thing on the ABI for ARM on linux, references to fp are replaced by sp the stack
pointer. Since this value is usually only meant to be collected there is no
impact on the tests.

Also, since some registers are not available on ARM some tests have been skipped
or adapted to work properly given that situation.

For example the pending.exp test checks if a trace frame is found, however since
some registers are not available some error output is generated about this issue
even if the trace frame is found. The test was made a little more lax so that
such output is accepted.

Other tests are left as failing since what they test is directly impacted by
those missing registers like reports.exp that will output errors like :
"Unable to access DWARF register number 8".

There is some known tests that are failing to due what seems to be a bug in the
ARM stack unwinder, for example collecting a string in collections.exp will not
work properly when calling tdump the first time since it can't compute the frame
id but will work when called the second time... Fixing this issue is left as
work to be done.

This issue also has ripple effects on some other tests like :
mi-trace-frame-collected.exp since it causes a "Register 13 is not available" to
be printed.

Other than that the patch was tested on Ubuntu 14.04 ARMv7 and x86 with no
regression.
With gdbserver-{native,extended} / { -marm -mthumb -marmv4t }

Note that a lot of tests are flaky on ARMv7 and care was taken to test each
flaky test before and after the patch to ensure no regression was found.

gdb/ChangeLog:
	* NEWS: Add news for tracepoins on ARM.

gdb/gdbserver/ChangeLog:
	* linux-arm-low.c (arm_supports_tracepoints): New function.
	(struct linux_target_ops): Add arm_supports_tracepoints operation.
	* linux-low.c (linux_supports_breakpoint_kinds): New function.
	* linux-low.h: Add supports_breakpoint_kinds operation.
	* mem-break.c (check_breakpoint_from_length): Adjust comment.
	(set_breakpoint_at_with_length): New function.
	* mem-break.h: Add set_breakpoint_at_with_length.
	* server.c (handle_query): Add TracepointKinds+.
	* target.h (struct target_ops): Add supports_breakpoint_kinds operation.
	* tracepoint.c (struct tracepoint): Add breakpoint_len field.
	(add_tracepoint): Add breakpoint_len init.
	(cmd_qtdp): Handle :K kind parameter.
	(install_tracepoint): Handle breakpoint_len.
	(cmd_qtstart): Likewise.
	* remote.c (remote_supports_tracepoint_kinds): Add
	TracepointKinds packet.
	(remote_download_tracepoint): Add breakpoint lenght to :K kind
	parameter.

gdb/testsuite/ChangeLog:
	* gdb.trace/backtrace.exp: Use sp for fp.
	* gdb.trace/collection.exp: Use sp for fp and sip tests that
	can't be supported on aarch32.
	* gdb.trace/mi-trace-frame-collected.exp:  Add aarch32 pc register.
	* gdb.trace/pending.exp: Be more permissive on test output in
	case of unrelated errors printed to stdout.
	* gdb.trace/report.exp: Use sp for fp.
	* gdb.trace/trace-break.exp: Add aarch32 sp register.
	* gdb.trace/while-dyn.exp: Likewise.

gdb/doc/ChangeLog:
	* doc/gdb.texinfo: Add documentation for QTDP kind parameter.