native-gdbserver: Clear isremote flag in board info
This patch clears the "isremote" flag in the native-gdbserver board.
The goal is to make is similar to the native-extended-gdbserver board,
and I believe that's the correct thing to do.

IIUC, DejaGnu automatically considers boards remote if their names don't
match the local hostname.  That means that native-gdbserver and
native-extended-gdbserver are considered remote by default by DejaGnu,
even though they run locally.  native-extended-gdbserver, however,
overrides its isremote flag to force it to be not remote.  So we are in
that weird state where native-gdbserver is considered remote, and
native-extended-gdbserver is considered not remote.

There seems to be some confusion globally in the testsuite about what's
native and what's not, what's remote and what's not.  It's probably
because of the overlap in vocabulary between DejaGnu and GDB.

From a DejaGnu point of view:

 - native: the host or target board is considered native if the its
   triplet is the same as the build system's triplet,
 - remote: the host or target board is considered remote if it's running
   on a different machine, and thus require ssh, for example, to run
   commands, versus simply running commands directly.

Note that they are not mutually exclusive, as you can have a remote
machine that has the same triplet as the build machine.

From a GDB point of view:

 - native: when GDB uses system calls such as ptrace to interact
   directly with processes on the same system its running on,
 - remote: when GDB speaks the Remote Protocol language with another
   program doing the ptrace stuff.

Note that they are mutually exclusive.  GDB can currently have only one
of them loaded at a specific time.

That means that there are cases where the target is !remote for DejaGnu,
but is remote for GDB (e.g. running gdbserver on the same machine).

You can also have a remote target for DejaGnu, but native for GDB (e.g.
building on x86 a GDB that runs on ARM and running the testsuite with
a remote host).

Therefore, a lot of tests in the testsuite don't check for the right
kind of remote.  For example, some use [is_remote target] (which checks
whether the DejaGnu target is remote), when what they really want to
know is whether GDB is using a remote target/the remote protocol,
because feature X is only available when GDB debugs natively.  Those
should use [gdb_is_target_remote] instead.

Another kind of tests checking the wrong thing is when they want to know
whether the current GDB target is able to, for example, attach a
process.  Some tests use [is_remote target] for this, since,
conveniently, native-extended-gdbserver is considered not remote and can
attach, while native-gdbserver is considered remote and can't attach.
What they ought to check is (I think, correct me if I'm wrong)
$use_gdb_stub.  That variable indicates if the remote gdb target we are
speaking to has the ability to run, attach/detach, etc.  It is set for
native-gdbserver and unset for native-extended-gdbserver.

So, flipping isremote off for native-gdbserver ends up breaking some
tests, or enables some tests that were previously skipped.  I believe
that it means that those tests are wrong in some way, and should be
improved.

I have started to fix some of them, but there are many left.  I'd still
like to submit this patch right now anyway, since I don't want to spend
countless hours fixing everything and then be told that my approach is
wrong... I'd like to know it right away :).  So I have included a few
examples of fixed tests in this patch, but it's by no means
comprehensive.