Transmit target symbol flags with qSymbol
When generating the fast tracepoints jump pad for ARM, we need to know
whether the gdb_collect symbol/function is a thumb one or an arm one.
Therefore, we need to communicate the target specific (or rather arch
specific?) flags found in GDB's struct symbol in the qSymbol response.

Those flags are set by the architecture on GDB's side, so should only be
meaningful to the corresponding target code in GDBserver.  The generic
GDBserver code transmits them, but does not understand them.

Retro-compatibility note: a newer GDB interacting with an older
GDBserver still works, since the older GDBserver only reads what it
needed (the value) and returned.  So more data won't affect it.  The new
GDBserver code also doesn't care about trailing data, so adding fields
in the same way should be possible in the future.

My understanding is that compatibility between an older GDB and a newer
GDBserver is not as much important, because it's easier to upgrade GDB,
compared to a GDBserver deployed in the field.  I still made sure that
that case is cleanly handled (i.e. doesn't crash), by issuing a warning.

gdb/gdbserver/ChangeLog:

	* remote-utils.c (struct sym_cache): Add target_flags field.
	(look_up_one_symbol): Add target_flags_ptr, refactor the code,
	parse target flags.  Move function comment to...
	* remote-utils.h (look_up_one_symbol): ... here.
	* symbol.c (find_minimal_symbol_address): Adjust call to
	look_up_one_symbol.
	* thread-db.c (thread_db_look_up_symbols): Likewise.
	* tracepoint.c (tracepoint_look_up_symbols): Likewise.

gdb/ChangeLog:

	* NEWS: Mention new sym_flags field in qSymbol packet.
	* remote.c (remote_check_symbols): Send symbol target flags with
	qSymbol answer.  Rename some variables.

gdb/doc/ChangeLog:

	* gdb.texinfo (General Query Packets): Add sym_flags field to
	qSymbol packet.