Record and propagate IPA symbols target flags to install_fast_tracepoint_jump_pad
In order to properly support ARM fast tracepoints, regardless if the IPA
is compiled in ARM or Thumb mode, we need to know if some functions
(such as gdb_collect) are ARM or Thumb, so we can generate the right
branch instruction.  Therefore, we need to save this information (the
target flags received from qSymbol) along with the symbol addresses in
the ipa_sym_addresses structure.  We then need to propagate this
information through the install_fast_tracepoint_jump_pad call chain.

This patch does that, including adapting the existing implementations.
There should be no user-visible changes with this patch.

gdb/gdbserver/ChangeLog:

	* linux-aarch64-low.c (aarch64_install_fast_tracepoint_jump_pad):
	Change collector and lockaddr type to struct ipa_symbol *.
	* linux-low.c (linux_install_fast_tracepoint_jump_pad):
	Likewise.
	* linux-low.h (linux_install_fast_tracepoint_jump_pad):
	Likewise.
	* linux-x86-low.c (amd64_install_fast_tracepoint_jump_pad,
	i386_install_fast_tracepoint_jump_pad,
	x86_install_fast_tracepoint_jump_pad): Likewise.
	* target.h (struct linux_target_ops):
	<install_fast_tracepoint_jump_pad>: Likewise.
	* tracepoint.c (struct ipa_sym_addresses): Change type of fields
	to struct ipa_symbol.
	(tracepoint_look_up_symbols): Record target flags of symbols.
	(install_fast_tracepoint): Adjust call to
	install_fast_tracepoint_jump_pad.
	(in_process_agent_supports_ust, clear_inferior_trace_buffer,
	get_jump_space_head, claim_trampoline_space,
	have_fast_tracepoint_trampoline_buffer, cmd_qtstart,
	stop_tracing, handle_tracepoint_bkpts,
	force_unlock_trace_buffer, fast_tracepoint_collecting,
	get_raw_reg_func_addr, get_get_tsv_func_addr,
	get_set_tsv_func_addr, target_malloc, download_tracepoint,
	download_trace_state_variables, upload_fast_traceframes): Adjust
	to type change of fields in struct ipa_sym_addresses.
	* tracepoint.h (struct ipa_symbol): New structure.