arm-tdep.c: Replace arguments to decode function by a structure
This patch prepares the code so that the "decode" functions can be moved
out of arm-tdep.c.  Instead of passing the arguments that are necessary
for the relocation directly through parameters, we now carry a pointer
to a data structure containing those values.  This allows decoupling the
decoding phase, which will be shared with gdbserver, from the relocation
phrase, which is gdb specific.