ftrace: Fix gdbserver crash when doing tstatus after detach or process exit
This patch fixes a gdbserver crash that is triggered by the following
sequence of events:

 - A process with the in-process agent loaded is debugged under gdbserver.
 - The process is detached or exits.
 - Commands tstatus or enable/disable with fast tracepoints are used.

Using either of tstatus or enable/disable ends up sending the qtstatus
packet to gdbserver.  During the handling of qtstatus, agent_loaded_p ()
returns true, even though the process that once had the agent loaded is
not present anymore.  We end up trying to read memory with
current_thread == NULL, causing a segfault here:

  gdb/gdbserver/linux-low.c:5560(linux_read_memory)[0x43583c]
  gdb/gdbserver/target.c:153(read_inferior_memory)[0x415d78]
  gdb/gdbserver/tracepoint.c:424(read_inferior_uinteger)[0x41c7fb]
  gdb/gdbserver/tracepoint.c:6288(upload_fast_traceframes)[0x425558]
  gdb/gdbserver/tracepoint.c:3645(cmd_qtstatus)[0x420dee]
  gdb/gdbserver/tracepoint.c:4239(handle_tracepoint_query)[0x4222ab]
  gdb/gdbserver/server.c:2543(handle_query)[0x411639]
  gdb/gdbserver/server.c:3910(process_serial_event)[0x413f39]
  gdb/gdbserver/server.c:4347(handle_serial_event)[0x415010]
  gdb/gdbserver/event-loop.c:428(handle_file_event)[0x41bed7]
  gdb/gdbserver/event-loop.c:184(process_event)[0x41b69e]
  gdb/gdbserver/event-loop.c:547(start_event_loop)[0x41c41d]
  gdb/gdbserver/server.c:3723(captured_main)[0x413a53]
  gdb/gdbserver/server.c:3802(main)[0x413c2f]

A first solution that comes to mind is to make agent_loaded_p check if
current_thread is NULL, and return false if that's the case.  It would
make sense, since if there is no current thread, the agent can't
possibly be loaded.  However, that would require adding some
 #ifdef GDBSERVER to the common code, which is not acceptable.

An alternative would be to use

  current_thread != NULL && agent_loaded_p ()

wherever agent_loaded_p () is used.  However, I find it error prone
for future uses of agent_loaded_p (), since it would be easy to forget
to check for current_thread.

Instead, the solution I chose is to clear the
all_agent_symbols_looked_up flag whenever we have no more current thread
(process exit or detach).  I am not 100% sure it's correct, as there
might be valid situations I don't know about where the agent is loaded
but current_thread == NULL, so please correct me if I'm wrong.

I added a test that either detaches from the process or makes it exit,
and then tries to use the commands that would cause the crash.  I run
the test both in all-stop and non-stop, since it was required to fix
both code paths in gdbserver.

Initially, I wanted to enhance gdb.trace/no-attach-trace.exp instead of
adding a new test case, since it tests a similar problem (gdbserver
would crash when doing tstart with no process attached).  However, my
case is specific to fast tracepoints and the in-process agent, whereas
no-attach-trace.exp also runs on targets that use normal tracing.  So I
left them as two separate test cases.

No regression with native-gdbserver && native-extended-gdbserver.

Finally, as a side-note, and just to make sure I understand correctly:
since there is a single global all_agent_symbols_looked_up flag, I guess
the tracking of whether the agent is loaded is not expected to work
correctly in a multi-process scenario, is that right?  If there are two
processes under gdbserver, there could be one with and one without the
agent.  So ideally (as it would be more "right" than the current patch),
I suppose we should track this per-process?

gdb/ChangeLog:

	* common/agent.h (agent_clear): New declaration.
	* common/agent.c (agent_clear): New function.

gdb/gdbserver/ChangeLog:

	* server.c (resume): Call agent_clear on inferior process exit.
	(process_serial_event): Call agent_clear on inferior process
	detach.
	(handle_target_event): Call agent_clear on inferior process
	exit.

gdb/testsuite/ChangeLog:

	* gdb.trace/ftrace-commands-after-detach-or-exit.exp: New file.
	* gdb.trace/ftrace-commands-after-detach-or-exit.c: New file.